Crate cvlcore

Modules

api

core

errors

ui

Constants

ANY_2_DIM_IMAGE

BGR_CV_IMAGE

Functions

calculate_mat_median

This method returns arithmetic mean (average) of all elements in array. In mathematics and statistics, the arithmetic mean / arithmetic average is the sum of a collection of numbers divided by the count of numbers in the collection. The collection is often a set of results from an experiment, an observational study, or a survey. The term “arithmetic mean” is preferred in some mathematics and statistics contexts because it helps distinguish it from other types of means, such as geometric and harmonic.

compute_vibration

This method returns image with vibrating pixels (colored by bounds values) by passed image. The main algorithm iterates over each pixel of Canny-image and calculate amount of nonzero pixels around current pixel. A target computed value replaced instead pixel value. The vibration image is network procedure which used for anxiety triggering by dispersion of statistic values.

gen_abs_frame

This recursive method returns result-image of opencv::absdiff() method by passed list of followed one by one frames of video stream. A result-image presents matrix Absolute difference between two 2D-arrays when they have the same size and type which used for removing from further analysis static pixels.

gen_abs_frame_reduce

This method returns reduced result-image of opencv::absdiff() method by passed list of followed one by one frames of video stream. A result-image presents matrix Absolute difference between two 2D-arrays when they have the same size and type which used for removing from further analysis static pixels.

gen_canny_frame

This method returns canny image from passed grayscale image by passed parameters. The Canny edge detector is an edge detection operator that uses a multi-stage algorithm to detect a wide range of edges in images. It was developed by John F. Canny in 1986. Canny also produced a computational theory of edge detection explaining why the technique works.

gen_canny_frame_by_sigma

This method returns canny image from passed grayscale image by passed parameters. The Canny edge detector is an edge detection operator that uses a multi-stage algorithm to detect a wide range of edges in images. It was developed by John F. Canny in 1986. Canny also produced a computational theory of edge detection explaining why the technique works.

gen_distribution_frame

This method returns distribution image from passed grayscale image by passed parameters. The distribution image is representation of the distribution of pixel gradients intensities in a digital image. As I mentioned in the introduction, image gradients are used as the basic building blocks in many computer vision and image processing applications. However, the network application of image gradients lies within edge detection.

gen_grayscale_frame

Transformations within RGB space like adding/removing the alpha channel, reversing the channel order, conversion to/from 16-bit RGB color (R5:G6:B5 or R5:G5:B5), as well as conversion to/from grayscale.

gen_threshold_frame

This method returns threshold image from passed bgr-image by passed black/white bounds values. The simplest thresholding methods replace each pixel in an image with a black pixel if the image intensity less than a fixed value called the threshold if the pixel intensity is greater than that threshold. This function is necessary for further image transformation to generate pixels vibration image.